Some cleanup and further removal of float operations.

+++ Grbl_Esp32/gcode.cpp
@@ -1086,9 +1086,9 @@ uint8_t gc_execute_line(char* line, uint8_t client) {
         if (gc_state.modal.spindle != SPINDLE_DISABLE) {
             if (bit_isfalse(gc_parser_flags, GC_PARSER_LASER_ISMOTION)) {
                 if (bit_istrue(gc_parser_flags, GC_PARSER_LASER_DISABLE))
-                    spindle->set_state(gc_state.modal.spindle, 0.0);
+                    spindle->set_state(gc_state.modal.spindle, 0);
                 else
-                    spindle->set_state(gc_state.modal.spindle, gc_block.values.s);
+                    spindle->set_state(gc_state.modal.spindle, (uint32_t)gc_block.values.s);
             }
         }
         gc_state.spindle_speed = gc_block.values.s; // Update spindle speed state.
@@ -1110,7 +1110,7 @@ uint8_t gc_execute_line(char* line, uint8_t client) {
         // Update spindle control and apply spindle speed when enabling it in this block.
         // NOTE: All spindle state changes are synced, even in laser mode. Also, pl_data,
         // rather than gc_state, is used to manage laser state for non-laser motions.
-        spindle->set_state(gc_block.modal.spindle, pl_data->spindle_speed);
+        spindle->set_state(gc_block.modal.spindle, (uint32_t)pl_data->spindle_speed);
         gc_state.modal.spindle = gc_block.modal.spindle;
     }
     pl_data->condition |= gc_state.modal.spindle; // Set condition flag for planner use.
@@ -1286,7 +1286,7 @@ uint8_t gc_execute_line(char* line, uint8_t client) {
                 if (!(settings_read_coord_data(gc_state.modal.coord_select, gc_state.coord_system)))
                     FAIL(STATUS_SETTING_READ_FAIL);
                 system_flag_wco_change(); // Set to refresh immediately just in case something altered.
-                spindle->set_state(SPINDLE_DISABLE, 0.0);
+                spindle->set_state(SPINDLE_DISABLE, 0);
                 coolant_set_state(COOLANT_DISABLE);
             }
             report_feedback_message(MESSAGE_PROGRAM_END);
